backend/pyproject.toml
[project]
name = "mediahelper-backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pyyaml",
    "sqlalchemy>=2.0.0",
    "alembic",
    "python-dotenv",
    "email-validator",
    "aiosqlite",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "loguru>=0.7.3",
    "bcrypt>=4.0.1"               
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
